requirements.txt
pandas
matplotlib
streamlit
hydralit_components
scikit-learn
scikit-surprise
plotly
